[INTERNAL] fix warnings (casting from int to size_t)
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.10@10655 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/ANALYSIS/OPENSWATH/MRMFeatureFinderScoring.h
@@ -355,7 +355,7 @@ private:
         " with " << transition_group.size()  << " nr transitions and nr chromats " << transition_group.getChromatograms().size() << std::endl;
 #endif
 
-        unsigned int group_size = transition_group.size();
+        int group_size = boost::numeric_cast<int>(transition_group.size());
         if (group_size == 0)
         {
           throw Exception::IllegalArgument(__FILE__, __LINE__, __PRETTY_FUNCTION__,
@@ -371,7 +371,7 @@ private:
         // calculate the normalized library intensity (expected value of the intensities)
         std::vector<double> normalized_library_intensity;
         transition_group.getLibraryIntensity(normalized_library_intensity);
-        OpenSwath::Scoring::normalize_sum(&normalized_library_intensity[0], normalized_library_intensity.size());
+        OpenSwath::Scoring::normalize_sum(&normalized_library_intensity[0], boost::numeric_cast<int>(normalized_library_intensity.size()));
 
         // calcxcorr -> for each lag do the correlation, normally use lag 0
         // xcorr_matrix  => correlate chromatogram i with chromatogram j
@@ -594,10 +594,10 @@ private:
     OpenSwath::SpectrumPtr getAddedSpectra_(OpenSwath::SpectrumAccessPtr swath_map, double RT, int nr_spectra_to_add)
     {
       std::vector<std::size_t> indices = swath_map->getSpectraByRT(RT, 0.0);
-      int closest_idx = indices[0];
+      int closest_idx = boost::numeric_cast<int>(indices[0]);
       if (indices[0] != 0 &&
-          std::fabs(swath_map->getSpectrumMetaById(indices[0] - 1).RT - RT) <
-          std::fabs(swath_map->getSpectrumMetaById(indices[0]).RT - RT))
+          std::fabs(swath_map->getSpectrumMetaById(boost::numeric_cast<int>(indices[0]) - 1).RT - RT) <
+          std::fabs(swath_map->getSpectrumMetaById(boost::numeric_cast<int>(indices[0])).RT - RT))
       {
         closest_idx--;
       }
