+++ source/ANALYSIS/OPENSWATH/DATAACCESS/DataAccessHelper.C
@@ -114,12 +114,11 @@ namespace OpenMS
             m.location = -1;
             m.unimod_id = rmod.getUniModAccession();
             p.modifications.push_back(m);
-
         }
         if ( !aa_sequence.getCTerminalModification().empty())
         {
             ResidueModification rmod = mod_db->getTerminalModification(aa_sequence.getCTerminalModification(), ResidueModification::C_TERM);
-            m.location = aa_sequence.size();
+            m.location = boost::numeric_cast<int>(aa_sequence.size());
             m.unimod_id = rmod.getUniModAccession();
             p.modifications.push_back(m);
         }
@@ -130,7 +129,7 @@ namespace OpenMS
             // search the residue in the modification database (if the sequence is valid, we should find it)
             ResidueModification rmod = mod_db->getModification(aa_sequence.getResidue(i).getOneLetterCode(),
                                                                aa_sequence.getResidue(i).getModification(), ResidueModification::ANYWHERE);
-            m.location = i;
+            m.location = boost::numeric_cast<int>(i);
             m.unimod_id = rmod.getUniModAccession();
             std::cout << " unimod accession " << m.unimod_id << std::endl;
             p.modifications.push_back(m);
@@ -161,7 +160,7 @@ namespace OpenMS
       aa_sequence = (String)peptide.sequence;
       for (std::vector<OpenSwath::LightModification>::const_iterator it = peptide.modifications.begin(); it != peptide.modifications.end(); ++it)
       {
-        TargetedExperimentHelper::setModification(it->location, peptide.sequence.size(), it->unimod_id, aa_sequence);
+        TargetedExperimentHelper::setModification(it->location, boost::numeric_cast<int>(peptide.sequence.size()), it->unimod_id, aa_sequence);
       }
   }
